added more ListIterator logic
